requirements.txt
celery
flower
dj-database-url
Django==5.2.9
django-ninja
django-sesame
httpx
numpy
openai
perplexityai
pgvector
pydantic
psycopg2-binary
python-crontab
python-dateutil
python-dotenv
redis
sqlparse
